# MVCCResolveWriteIntent{,Range} optimizations.