@petermattis 2016-02-16
Note that despite `Delete100` seeing the bulk of the gains from this change, the issue in #4386 is still present. (I had a vague hope that this would help).