@bdarnell 2016-02-16
LGTM

---

Review status: 0 of 2 files reviewed at latest revision, 1 unresolved discussion.

---

*[storage/engine/mvcc.go, line 1362 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4391#-KAfJtAvmfJGGAVyOEPx:-KAfJtAvmfJGGAVyOEPy:1582641414) ([raw file](https://github.com/cockroachdb/cockroach/blob/dccc190f7e12a5530fa7af13f009881dfc9a3007/storage/engine/mvcc.go#L1362)):*
Should we also un-defer `iter.Close()` just above?

---


*Comments from the [review on Reviewable.io](https://reviewable.io:443/reviews/cockroachdb/cockroach/4391#-:-KAfKJ9Gxr1Zh86QvcXg:676834233)*
<!-- Sent from Reviewable.io -->