@tschottdorf 2016-02-16
Reviewed 2 of 2 files at r1.
Review status: all files reviewed at latest revision, 5 unresolved discussions.

---

*[storage/engine/mvcc.go, line 1434 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4391#-KAfN3syf-A-ZvhTafgz:-KAfN3syf-A-ZvhTafh-:-1121455310) ([raw file](https://github.com/cockroachdb/cockroach/blob/dccc190f7e12a5530fa7af13f009881dfc9a3007/storage/engine/mvcc.go#L1434)):*
```go
newMeta.Txn = &buf.newTxn
*newMeta.Txn = intent.Txn
```

---

*[storage/engine/mvcc.go, line 1586 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4391#-KAfOIf3mZRx8wXbobZ5:-KAfOIf3mZRx8wXbobZ6:-1413949307) ([raw file](https://github.com/cockroachdb/cockroach/blob/dccc190f7e12a5530fa7af13f009881dfc9a3007/storage/engine/mvcc.go#L1586)):*
s/that/the/

---

*[storage/engine/mvcc.go, line 1587 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4391#-KAfNs1cFbfpfL1kTJ2f:-KAfNs1cFbfpfL1kTJ2g:-2066982994) ([raw file](https://github.com/cockroachdb/cockroach/blob/dccc190f7e12a5530fa7af13f009881dfc9a3007/storage/engine/mvcc.go#L1587)):*
s/roacp/roach/, and I think you should have a helper next to `key.Next()` which contains the meat of `key.Next()` but is more efficient.

---

*[storage/engine/mvcc.go, line 1589 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4391#-KAfOQC3taNejjaksfuX:-KAfOQC3taNejjaksfuY:721949179) ([raw file](https://github.com/cockroachdb/cockroach/blob/dccc190f7e12a5530fa7af13f009881dfc9a3007/storage/engine/mvcc.go#L1589)):*
s/append\((.*)\)/\1/?

---


*Comments from the [review on Reviewable.io](https://reviewable.io:443/reviews/cockroachdb/cockroach/4391)*
<!-- Sent from Reviewable.io -->