@tamird 2016-02-16
Reviewed 1 of 2 files at r1.
Review status: all files reviewed at latest revision, 9 unresolved discussions.

---

*[storage/engine/mvcc.go, line 1423 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4391#-KAbrxYEI2eSJyxsH1NL:-KAbrxYEI2eSJyxsH1NM:-1524201924) ([raw file](https://github.com/cockroachdb/cockroach/blob/dccc190f7e12a5530fa7af13f009881dfc9a3007/storage/engine/mvcc.go#L1423)):*
is it me, or is this just an odd way to write
```
buf.newMeta = *meta
newMeta := &buf.newMeta
```
Could you get rid of newMeta entirely?

---

*[storage/engine/mvcc.go, line 1432 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4391#-KAbsDIWPKWzY_WJFrgG:-KAbsDIWPKWzY_WJFrgH:-396323780) ([raw file](https://github.com/cockroachdb/cockroach/blob/dccc190f7e12a5530fa7af13f009881dfc9a3007/storage/engine/mvcc.go#L1432)):*
ditto
```
buf.newTxn = intent.Txn
newMeta.Txn = &buf.newTxn
```

---

*[storage/engine/mvcc.go, line 1585 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4391#-KAbtc-FqSuxVf2f8SiA:-KAbtc-FqSuxVf2f8SiB:278024818) ([raw file](https://github.com/cockroachdb/cockroach/blob/dccc190f7e12a5530fa7af13f009881dfc9a3007/storage/engine/mvcc.go#L1585)):*
the copy is happening pretty far from here

---

*[storage/engine/mvcc.go, line 1589 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4391#-KAbtSzrmyPK_EwJwj3I:-KAbtSzrmyPK_EwJwj3J:1726010711) ([raw file](https://github.com/cockroachdb/cockroach/blob/dccc190f7e12a5530fa7af13f009881dfc9a3007/storage/engine/mvcc.go#L1589)):*
what does this construction do?

---


*Comments from the [review on Reviewable.io](https://reviewable.io:443/reviews/cockroachdb/cockroach/4391)*
<!-- Sent from Reviewable.io -->