@tamird 2016-02-16
Reviewed 1 of 2 files at r1.
Review status: all files reviewed at latest revision, 9 unresolved discussions.

---



*Comments from the [review on Reviewable.io](https://reviewable.io:443/reviews/cockroachdb/cockroach/4391)*
<!-- Sent from Reviewable.io -->