@petermattis 2016-02-16
Review status: 1 of 2 files reviewed at latest revision, 9 unresolved discussions.

---

*[storage/engine/mvcc.go, line 1362 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4391#-KAfJtAvmfJGGAVyOEPx:-KAfhukiXuzmVWTRIpkG:-750253025) ([raw file](https://github.com/cockroachdb/cockroach/blob/dccc190f7e12a5530fa7af13f009881dfc9a3007/storage/engine/mvcc.go#L1362)):*
Defers are not actually slow. I think the measurable slowness comes from a capturing closure such as `putBufferPool.Put(buf)` vs a non-capturing closure like `iter.Close()`. I need to investigate this more at some point, but the short answer is that it doesn't make a difference to not defer `iter.Close()`.

---

*[storage/engine/mvcc.go, line 1423 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4391#-KAbrxYEI2eSJyxsH1NL:-KAfiEMISNTjdPc0Gzyo:-817474413) ([raw file](https://github.com/cockroachdb/cockroach/blob/dccc190f7e12a5530fa7af13f009881dfc9a3007/storage/engine/mvcc.go#L1423)):*
Yes, I could get rid of the `newMeta` entirely. I don't find the code as written odd, but then again, I wrote it. 

---

*[storage/engine/mvcc.go, line 1432 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4391#-KAbsDIWPKWzY_WJFrgG:-KAfktDqt18-_C5TrelD:-499146009) ([raw file](https://github.com/cockroachdb/cockroach/blob/dccc190f7e12a5530fa7af13f009881dfc9a3007/storage/engine/mvcc.go#L1432)):*
Done.

---

*[storage/engine/mvcc.go, line 1434 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4391#-KAfN3syf-A-ZvhTafgz:-KAfkxYtqCjcfpTKpjMP:-499146009) ([raw file](https://github.com/cockroachdb/cockroach/blob/dccc190f7e12a5530fa7af13f009881dfc9a3007/storage/engine/mvcc.go#L1434)):*
Done.

---

*[storage/engine/mvcc.go, line 1585 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4391#-KAbtc-FqSuxVf2f8SiA:-KAfl7ZiCFohp7mL28d4:596435777) ([raw file](https://github.com/cockroachdb/cockroach/blob/dccc190f7e12a5530fa7af13f009881dfc9a3007/storage/engine/mvcc.go#L1585)):*
Yeah, this comment made more sense in an early iteration of this PR. Will remove (there is already a comment next to where the copy happens).

---

*[storage/engine/mvcc.go, line 1586 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4391#-KAfOIf3mZRx8wXbobZ5:-KAfl46I_QtMX4a8gYSP:-499146009) ([raw file](https://github.com/cockroachdb/cockroach/blob/dccc190f7e12a5530fa7af13f009881dfc9a3007/storage/engine/mvcc.go#L1586)):*
Done.

---

*[storage/engine/mvcc.go, line 1587 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4391#-KAfNs1cFbfpfL1kTJ2f:-KAflLfd0eebxLruVOl8:-1990792654) ([raw file](https://github.com/cockroachdb/cockroach/blob/dccc190f7e12a5530fa7af13f009881dfc9a3007/storage/engine/mvcc.go#L1587)):*
My fingers have trouble spelling `roachpb`. Do you have a

---

*[storage/engine/mvcc.go, line 1589 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4391#-KAbtSzrmyPK_EwJwj3I:-KAflT7maCzBUrQ_wwWF:715378697) ([raw file](https://github.com/cockroachdb/cockroach/blob/dccc190f7e12a5530fa7af13f009881dfc9a3007/storage/engine/mvcc.go#L1589)):*
Nothing. I have no idea where that came from.

---


*Comments from the [review on Reviewable.io](https://reviewable.io:443/reviews/cockroachdb/cockroach/4391)*
<!-- Sent from Reviewable.io -->